Initial template for USOS requests created
