--- ElkaUWP.DataLayer/Usos/Repositories/BuildingReadOnlyRepository.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using ElkaUWP.DataLayer.Usos.Entities;
-
-namespace ElkaUWP.DataLayer.Usos.Repositories
-{
-    public class BuildingReadOnlyRepository
-    {
-        public List<Building> GetBuildings()
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
