+++ ElkaUWP.Infrastructure/ElkaUWP.Infrastructure.csproj
@@ -130,7 +130,6 @@
     <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Abstractions\Bases\OAuthProtectedResourceRequestWrapperBase.cs" />
     <Compile Include="Abstractions\Interfaces\IUsosOAuthService.cs" />
     <Compile Include="Constants.App.cs" />
     <Compile Include="Constants.Usos.cs" />
