+++ ElkaUWP.Infrastructure/Extensions/SettingsStorageExtensions.cs
@@ -58,7 +58,7 @@ namespace ElkaUWP.Infrastructure.Extensions
 
         public static async Task<StorageFile> SaveFileAsync(this StorageFolder folder, byte[] content, string fileName, CreationCollisionOption options = CreationCollisionOption.ReplaceExisting)
         {
-            if (content == null)
+            if (content is null)
             {
                 throw new ArgumentNullException(paramName: nameof(content));
             }
@@ -77,7 +77,7 @@ namespace ElkaUWP.Infrastructure.Extensions
         {
             var item = await folder.TryGetItemAsync(name: fileName).AsTask().ConfigureAwait(false);
 
-            if ((item == null) || !item.IsOfType(type: StorageItemTypes.File))
+            if ((item is null) || !item.IsOfType(type: StorageItemTypes.File))
                 return null;
 
             var storageFile = await folder.GetFileAsync(name: fileName);
